% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{print.pm_params}
\alias{print.pm_params}
\title{Method to print \code{pm_params} object}
\usage{
\method{print}{pm_params}(x, ...)
}
\arguments{
\item{x}{\code{pm_params} object}

\item{...}{Ignored}
}
\value{
\code{x}, invisibly.
}
\description{
Method to print \code{pm_params} object
}
\examples{
para <- pm_params()
para
}
\keyword{internal}
