% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_metrics.R
\name{assess_risk_diff}
\alias{assess_risk_diff}
\title{Compute the upper bound of the differencing risk, following the proposition 7
of the paper.}
\usage{
assess_risk_diff(sigma_eps, beta)
}
\arguments{
\item{sigma_eps}{double}

\item{beta}{double}
}
\value{
double vector
}
\description{
Compute the upper bound of the differencing risk, following the proposition 7
of the paper.
}
\examples{
assess_risk_diff(sigma_eps = 0.1, beta = 0.2)
}
