% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration_differencing.R
\name{pm_calib_diff}
\alias{pm_calib_diff}
\title{Calibration of the differencing noise (sigma_epsilon)}
\usage{
pm_calib_diff(params = NULL, beta = NULL, tau = NULL, level = 0.95)
}
\arguments{
\item{params}{object pm_params}

\item{beta}{real vector of beta values (accuracy level of the inference)}

\item{tau}{real vector of tau values (ceiling risk level)}

\item{level}{confidence level for the confidence intervall of the relative loss}
}
\value{
data.frame and pm_calib_diff object
}
\description{
Calibration of the differencing noise (sigma_epsilon)
}
\details{
By default, the function returns the results for some beta and tau values
beta in (0.05, 0.1, 0.15, 0.2, 0.25) and
tau in s(0.5, 0.55, 0.6, 0.65, ..., 0.90, 0.95)
}
\examples{
# default behaviour
pm_calib_diff()
}
