% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{print.pm_params}
\alias{print.pm_params}
\title{Method to print pm_params object}
\usage{
\method{print}{pm_params}(x, ...)
}
\arguments{
\item{x}{pm_params object}

\item{...}{Ignored}
}
\value{
pm_params object x
}
\description{
Method to print pm_params object
}
\examples{
para <- pm_params()
para
}
